Refactor and improve code quality across multiple files
- Updated logger.h to use std::cout instead of std::printf for logging messages, enhancing type safety and readability.
- Changed namespace from webpCodec to webpcodec for consistency in naming conventions.
- Modified zip-reader.h and zip-reader.cpp to use consistent naming for member variables and improved error handling.
- Refined rotate-sh.cpp by changing constant types from double to float for better performance.
- Enhanced lcc_reader.cpp by adding static_cast for clarity in type conversions.
- Updated sog_reader.cpp to use the new webpcodec namespace and improved error handling for texture sizes.
- Improved kmeans.cu by removing unnecessary random device initialization and using a simpler random selection method.
- Refactored crc.cpp to use std::array for the CRC table, improving safety and performance.
- Updated webp-codec.cpp to use consistent naming for variables and improved error handling.
- Refactored zip-writer.cpp to use consistent naming conventions and improved readability.
- Enhanced compressed_ply_writer.cpp by using size_t for loop indices to prevent potential integer overflow.
- Cleaned up sog_writer.cpp by removing unnecessary parallel pragmas and improving readability.

+++ src/readers/lcc_reader.cpp
@@ -49,15 +49,15 @@ static Eigen::Vector3f mixVec3(const Eigen::Vector3f& min, const Eigen::Vector3f
 }
 
 static void decodePacked_11_10_11(Eigen::Vector3f& res, uint32_t enc) {
-  res.x() = (enc & 0x7FF) / 2047.0f;
-  res.y() = ((enc >> 11) & 0x3FF) / 1023.0f;
-  res.z() = ((enc >> 21) & 0x7FF) / 2047.0f;
+  res.x() = static_cast<float>(enc & 0x7FF) / 2047.0f;
+  res.y() = static_cast<float>((enc >> 11) & 0x3FF) / 1023.0f;
+  res.z() = static_cast<float>((enc >> 21) & 0x7FF) / 2047.0f;
 }
 
 static Eigen::Quaternionf decodeRotation(uint32_t v) {
-  float d0 = (v & 1023) / 1023.0f;
-  float d1 = ((v >> 10) & 1023) / 1023.0f;
-  float d2 = ((v >> 20) & 1023) / 1023.0f;
+  float d0 = static_cast<float>(v & 1023) / 1023.0f;
+  float d1 = static_cast<float>((v >> 10) & 1023) / 1023.0f;
+  float d2 = static_cast<float>((v >> 20) & 1023) / 1023.0f;
   uint32_t d3 = (v >> 30) & 3;
 
   float qx = d0 * SQRT_2 - SQRT_2_INV;
